requirements.txt
pyyaml
requests
PyTelegramBotAPI